server/requirements.txt
sentence-transformers>=2.2.0
fastapi>=0.100.0
uvicorn>=0.23.0
psycopg[binary]>=3.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
torch>=2.0.0
